Added Windows Support (#1)
* Add build script for configuring and building ParticleZoo on Windows with MSVC

* Fixed some minor issues that trigger build warnings on MSVC

* Updated README file with Windows build and installation instructions

+++ src/egs/egsphspFile.cc
@@ -167,7 +167,7 @@ namespace ParticleZoo::EGSphspFile
 
         std::uint64_t historiesRecorded = getHistoriesWritten();
         if (historiesRecorded > numberOfOriginalHistories_) {
-            numberOfOriginalHistories_ = historiesRecorded;
+            numberOfOriginalHistories_ = static_cast<float>(historiesRecorded);
         }
 
         buffer.write<float>(numberOfOriginalHistories_);
